+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `go-huggingface` Changelog
 
+## Next
+
+- Package `datasets`:
+  - Added `datasets` package for downloading and iterating over parquet files of datasets from the HuggingFace Hub.
+  - Added `cmd/generate_dataset_structs` for generating Go structs for dataset records.
+
 ## v0.3.4
 
 - Added `models/transformer` package for loading HuggingFace transformer models (Experimental).